+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(ncdf4)
+
+test_check("ncdf4")