+++ NEWS.md
@@ -1,5 +1,7 @@
 # ncdf4 1.16.9001
 
+* Now works for Thredds sources on Windows (only 64-bit). 
+
 * Added ROpenSci winlibs build process.
 
 * Forked from CRAN ncdf4 1.16. (mdsumner@gmail.com 2018-04-17)